refactor: utils/parse_sse_stream.ts 파서 로직 단독 보관 (#266)
